Cast address to SIZE_T before casting to PTR_ULONG in src/inc/clrnt.h

+++ src/inc/clrnt.h
@@ -916,7 +916,7 @@ RtlpGetFunctionEndAddress (
     if ((FunctionLength & 3) != 0) {
         FunctionLength = (FunctionLength >> 2) & 0x7ff;
     } else {
-        FunctionLength = *(PTR_ULONG)(ImageBase + FunctionLength) & 0x3ffff;
+        FunctionLength = *(PTR_ULONG)((SIZE_T)(ImageBase + FunctionLength)) & 0x3ffff;
     }
     
     return FunctionEntry->BeginAddress + 2 * FunctionLength;
